+++ src/app/Wallets/SelectedWallet/SelectedWallet.tsx
@@ -1,13 +1,13 @@
 import { Unplug, Wallet } from "lucide-react";
-import { useWalletProvider } from "../hooks/useWalletProvider";
-import { Button } from "./ui/button";
+import { Button } from "@/components/ui/button";
 import {
   Card,
   CardContent,
   CardDescription,
   CardHeader,
   CardTitle,
-} from "./ui/card";
+} from "@/components/ui/card";
+import { useWalletProvider } from "@/hooks/useWalletProvider";
 
 export const SelectedWallet = () => {
   const { selectedWallet, selectedAccount, disconnectWallet } =
